+++ app/src/androidTest/java/woowacourse/payments/NumberTextFieldTest.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTextInput
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.component.NumberTextField
+import woowacourse.payments.ui.component.NumberTextField
 
 class NumberTextFieldTest {
     @get:Rule
