+++ app/src/androidTest/java/woowacourse/payments/StringTextFieldTest.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTextInput
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.component.StringTextField
+import woowacourse.payments.ui.component.StringTextField
 
 class StringTextFieldTest {
     @get:Rule
