+++ app/src/main/java/woowacourse/payments/AddPaymentCardScreen.kt
@@ -16,10 +16,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import woowacourse.payments.component.NewCardTopBar
-import woowacourse.payments.component.NumberTextField
-import woowacourse.payments.component.PaymentCard
-import woowacourse.payments.component.StringTextField
+import woowacourse.payments.ui.component.NewCardTopBar
+import woowacourse.payments.ui.component.NumberTextField
+import woowacourse.payments.ui.component.PaymentCard
+import woowacourse.payments.ui.component.StringTextField
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
 @Composable
